- fixed an issue with overlapping reservations not being counted correctly towards capacity - displaying earliest time capacity is reached when displaying policy problems
